rwengine: Implement actual braking when the vehicle is going forward
This implements proper braking by detecting a brake condition with a
negative throttle and a positive forward velocity. Under this condition,
the engine is stopped and brake is applied. A mass coefficient is
introduced to quantify how heavy a vehicle is, so that more braking can
be applied to heavy vehicles, that have higher inertia.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>