rwengine: Set the linear velocity to a zero vector when stopped
Vehicles have a tendency to move on their own when they are stopped, due
to their interactions with other surfaces that result in a non-zero
velocity.

Since we have a threshold with the isStopped() method, fix this situation
by setting the linear velocity and wheels rotation (that otherwise also
rotate on their own) to zero vectors.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>